# improve performance of MHD (quite a bit) and Euler (a little bit)