I know, you will hate me that I didn't make these changes earlier, but I was just motivated by the first performance comparisons you showed earlier today... 

I get the following results running `julia --check-bounds=no --threads=1` on `main`.
```julia
julia> using Trixi, BenchmarkTools

julia> begin # performance
           elixirs = [
               joinpath(examples_dir(), "structured_3d_dgsem", "elixir_mhd_alfven_wave.jl"),]
           tspan = (0.0, 1.0e-5)
           
           for elixir in elixirs
               println(basename(elixir))
               redirect_stdout(devnull) do
                   trixi_include(elixir, tspan=tspan)
               end
               u_ode = copy(sol.u[end]); du_ode = similar(u_ode)
               display(@benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0))
               println()
           end
       end
elixir_mhd_alfven_wave.jl
BenchmarkTools.Trial: 1217 samples with 1 evaluation.
 Range (min … max):  4.035 ms …  4.638 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.088 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.106 ms ± 71.805 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅    █▂▁▂ ▅▅ ▁ ▁▃    ▁▁                                     
  █▁▄▄▁████████████▇▇█▇██▇▇▇█▇▆▅▅▇▅▅▅▆▆▅▅▅▆▅▅▅▅▇▇▅▅▅▆▅▆▅▅▅▆▆ █
  440 ms       Histogram: log(frequency) by time     4.35 ms <

 Memory estimate: 480 bytes, allocs estimate: 5.
```
With this PR, I get
```julia
julia> using Trixi, BenchmarkTools

julia> begin # performance
           elixirs = [
               joinpath(examples_dir(), "structured_3d_dgsem", "elixir_mhd_alfven_wave.jl"),]
           tspan = (0.0, 1.0e-5)
           
           for elixir in elixirs
               println(basename(elixir))
               redirect_stdout(devnull) do
                   trixi_include(elixir, tspan=tspan)
               end
               u_ode = copy(sol.u[end]); du_ode = similar(u_ode)
               display(@benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0))
               println()
           end
       end
elixir_mhd_alfven_wave.jl
BenchmarkTools.Trial: 1428 samples with 1 evaluation.
 Range (min … max):  3.416 ms …  4.637 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.474 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.501 ms ± 79.232 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅    ▁ ▂█▁▁ ▆▁ ▁▂▂                                          
  █▄▆▁▇██████▇██████▆████▇▆▆▆▇▇▇▅▆▆█▆▆▆▄▄▅▄▁▅▄▁▄▄▄▅▄▄▄▄▆▁▄▁▄ █
  3.42 ms      Histogram: log(frequency) by time     3.81 ms <

 Memory estimate: 480 bytes, allocs estimate: 5.
```
That's a speed-up of ca. 15%. We can get an additional speed-up of ca. 3% once https://github.com/JuliaArrays/StaticArrays.jl/pull/949 is merged (but we can ignore this for our performance comparison).