@ranocha 2021-08-05
We will have to wait a bit for CI since GitHub actions are partially down, see https://www.githubstatus.com/incidents/7p1nnvkgh96y